Add Printf/Gnl v2

+++ includes/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: arpenel <arpenel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/13 07:03:03 by arpenel           #+#    #+#             */
-/*   Updated: 2025/02/15 13:01:44 by arpenel          ###   ########.fr       */
+/*   Updated: 2025/04/10 18:29:09 by arpenel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,6 +29,6 @@ char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strdup(const char *s);
 char	*ft_strchr(const char *s, int c);
 char	*ft_strjoin(char const *s1, char const *s2);
-int		ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
 
 #endif
